+++ CHANGELOG.md
@@ -9,9 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added support for sections/indexes to directory accounts APIs [#5701](https://github.com/rokwire/illinois-app/issues/5701).
 - Allow custom image auth headers for triangle header bar [#5863](https://github.com/rokwire/illinois-app/issues/5863).
+- Added ListEx extension helper [#5873](https://github.com/rokwire/illinois-app/issues/5873).
+- Added StringUtils.containsHtmlTags helper [#5873](https://github.com/rokwire/illinois-app/issues/5873).
 ### Fixed
 - Fixed crash in ExpansionTile widget [#5969](https://github.com/rokwire/illinois-app/issues/5969).
 - Fixed crash when truncating strings [#5988](https://github.com/rokwire/illinois-app/issues/5988).
+- Do not log images exceptions as fatal errors [#722](https://github.com/rokwire/app-flutter-plugin/issues/722).
 
 ## [1.13.5] - 2026-08-11
 ### Added